+++ include/listopad/performance_log.h
@@ -101,6 +101,8 @@ enum class PerformanceLogFilter {
   Processor,
   Memory,
   Disk,
+  SqlServer,
+  Processes,
 };
 
 struct PerformanceLogUiState {